requirements.txt
PyJWT
flask
flask_restful
cryptography
